+++ example/lib/db/driftdb.g.dart
@@ -262,9 +262,10 @@ class TodosCompanion extends UpdateCompanion<Todo> {
 }
 
 class $TodosTable extends Todos with TableInfo<$TodosTable, Todo> {
-  final GeneratedDatabase _db;
+  @override
+  final GeneratedDatabase attachedDatabase;
   final String? _alias;
-  $TodosTable(this._db, [this._alias]);
+  $TodosTable(this.attachedDatabase, [this._alias]);
   final VerificationMeta _idMeta = const VerificationMeta('id');
   @override
   late final GeneratedColumn<int?> id = GeneratedColumn<int?>(
@@ -383,7 +384,7 @@ class $TodosTable extends Todos with TableInfo<$TodosTable, Todo> {
 
   @override
   $TodosTable createAlias(String alias) {
-    return $TodosTable(_db, alias);
+    return $TodosTable(attachedDatabase, alias);
   }
 }
 
@@ -643,9 +644,10 @@ class ATodoDontCompanion extends UpdateCompanion<ATodoDontData> {
 
 class $ATodoDontTable extends ATodoDont
     with TableInfo<$ATodoDontTable, ATodoDontData> {
-  final GeneratedDatabase _db;
+  @override
+  final GeneratedDatabase attachedDatabase;
   final String? _alias;
-  $ATodoDontTable(this._db, [this._alias]);
+  $ATodoDontTable(this.attachedDatabase, [this._alias]);
   final VerificationMeta _idMeta = const VerificationMeta('id');
   @override
   late final GeneratedColumn<int?> id = GeneratedColumn<int?>(
@@ -764,7 +766,7 @@ class $ATodoDontTable extends ATodoDont
 
   @override
   $ATodoDontTable createAlias(String alias) {
-    return $ATodoDontTable(_db, alias);
+    return $ATodoDontTable(attachedDatabase, alias);
   }
 }